propositions film de la search bar
